Move useCachedDomainContactDetails to own file
